Move homeworks base entity to separate module (#126097)
* Move homeworks base entity to separate module

* Move calculate_unique_id to util.py
