+++ homeassistant/components/homeworks/binary_sensor.py
@@ -15,7 +15,7 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import HomeworksData, HomeworksEntity, HomeworksKeypad
+from . import HomeworksData, HomeworksKeypad
 from .const import (
     CONF_ADDR,
     CONF_BUTTONS,
@@ -25,6 +25,7 @@ from .const import (
     CONF_NUMBER,
     DOMAIN,
 )
+from .entity import HomeworksEntity
 
 _LOGGER = logging.getLogger(__name__)
