+++ homeassistant/components/homeworks/button.py
@@ -13,7 +13,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import HomeworksData, HomeworksEntity
+from . import HomeworksData
 from .const import (
     CONF_ADDR,
     CONF_BUTTONS,
@@ -23,6 +23,7 @@ from .const import (
     CONF_RELEASE_DELAY,
     DOMAIN,
 )
+from .entity import HomeworksEntity
 
 
 async def async_setup_entry(
